requirements.txt
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
ragas>=0.1.14
datasets>=2.19.0
groq>=0.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0